+++ e2e/sinon/nestjs/automock-sinon-ctor-injection.e2e.test.ts
@@ -1,3 +1,5 @@
+import 'reflect-metadata';
+
 import { UnitReference } from '@automock/core';
 import { TestBed } from '@automock/sinon';
 import {