+++ packages/testbeds/sinon/src/index.ts
@@ -1,7 +1,9 @@
 /// <reference types="@types/sinon" />
 
 import { Type } from '@automock/types';
+import { IdentifierMetadata } from '@automock/common';
 import { SinonStubbedInstance } from 'sinon';
+
 export * from './testbed-factory';
 
 declare module '@automock/core' {
@@ -20,7 +22,8 @@ declare module '@automock/core' {
    */
   export interface UnitReference {
     /**
-     * Retrieves a reference to the mocked object of a dependency corresponding to its type identifier.
+     * Retrieves a reference to the mocked object of a dependency corresponding to its type
+     * identifier.
      *
      * @template TDependency The type of the dependency being retrieved.
      * @param type The type representing the dependency.
@@ -29,37 +32,93 @@ declare module '@automock/core' {
     get<TDependency>(type: Type<TDependency>): SinonStubbedInstance<TDependency>;
 
     /**
-     * Retrieves a reference to the mocked object of a dependency corresponding to a string-based token.
+     * Retrieves a reference to the mocked object of a dependency corresponding to its
+     * type identifier and the identifier metadata.
+     *
+     * @since 2.1.0
+     * @template TDependency The type of the dependency being retrieved.
+     * @param type The type representing the dependency
+     * @param identifierMetadata An accompanying metadata object for the token identifier.
+     * @returns SinonStubbedInstance<TDependency> The mocked object corresponding to the provided
+     * type identifier and identifier metadata.
+     */
+    get<TDependency>(
+      type: Type<TDependency>,
+      identifierMetadata: IdentifierMetadata
+    ): SinonStubbedInstance<TDependency>;
+
+    /**
+     * Retrieves a reference to the mocked object of a dependency corresponding to a string-based
+     * token.
      *
      * @template TDependency The type of the dependency being retrieved.
      * @param token The string-based token representing the dependency.
-     * @returns The mocked object corresponding to the provided string-based token.
+     * @returns SinonStubbedInstance<TDependency> The mocked object corresponding to the provided
+     * string-based token.
      */
     get<TDependency>(token: string): SinonStubbedInstance<TDependency>;
 
     /**
-     * Retrieves a reference to the mocked object of a dependency corresponding to a symbol-based token.
+     * Retrieves a reference to the mocked object of a dependency corresponding to its
+     * string-based identifier and the identifier metadata.
+     *
+     * @since 2.1.0
+     * @template TDependency The type of the dependency being retrieved.
+     * @param token The string-based token representing the dependency.
+     * @param identifierMetadata An accompanying metadata object for the token identifier.
+     * @returns SinonStubbedInstance<TDependency> The mocked object corresponding to the provided
+     * string-based token and identifier metadata.
+     */
+    get<TDependency>(
+      token: string,
+      identifierMetadata: IdentifierMetadata
+    ): SinonStubbedInstance<TDependency>;
+
+    /**
+     * Retrieves a reference to the mocked object of a dependency corresponding to a symbol-based
+     * token.
      *
      * @since 2.0.0
      * @template TDependency The type of the dependency being retrieved.
      * @param token The symbol-based token representing the dependency.
-     * @returns The mocked object corresponding to the provided symbol-based token.
+     * @returns SinonStubbedInstance<TDependency> The mocked object corresponding to the provided
+     * symbol-based token.
      */
     get<TDependency>(token: symbol): SinonStubbedInstance<TDependency>;
 
     /**
-     * Retrieves a mocked object or a constant value of a dependency using its type, string, or symbol token.
+     * Retrieves a reference to the mocked object of a dependency corresponding to its
+     * symbol-based identifier and the identifier metadata.
      *
-     * This method provides flexibility in retrieving dependencies by allowing various identifier types.
-     * Depending on the identifier and the setup, it can return either a mocked object or a constant value.
-     *
-     * @since 2.0.0
+     * @since 2.1.0
      * @template TDependency The type of the dependency being retrieved.
-     * @param identifier The type or token that the dependency corresponds to.
-     * @returns The mocked object corresponding to the provided identifier.
+     * @param token The symbol-based token representing the dependency.
+     * @param identifierMetadata An accompanying metadata object for the token identifier.
+     * @returns SinonStubbedInstance<TDependency> The mocked object corresponding to the provided
+     * symbol-based token and identifier metadata.
      */
     get<TDependency>(
-      identifier: Type<TDependency> | string | symbol
+      token: symbol,
+      identifierMetadata: IdentifierMetadata
+    ): SinonStubbedInstance<TDependency>;
+
+    /**
+     * Retrieves a reference to the mocked object of a dependency corresponding to its
+     * type, string-based or symbol-based identifier and the identifier metadata if present.
+     *
+     * This method provides flexibility in retrieving dependencies by allowing various identifier
+     * types.
+     *
+     * @since 2.1.0
+     * @template TDependency The type of the dependency being retrieved.
+     * @param identifier The type or token that the dependency corresponds to.
+     * @param identifierMetadata
+     * @returns SinonStubbedInstance<TDependency> The mocked object corresponding to the provided
+     * identifier, along with any available identifier metadata.
+     */
+    get<TDependency>(
+      identifier: Type<TDependency> | string | symbol,
+      identifierMetadata?: IdentifierMetadata
     ): SinonStubbedInstance<TDependency>;
   }
 }